requirements.txt
# Core dependencies
python-dotenv==1.0.0
PyYAML==6.0.1

# Reddit API
praw==7.7.1

# YouTube API
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1

# Ollama
ollama==0.4.4

# Database
aiosqlite==0.19.0

# Scheduling
schedule==1.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Text processing
# tiktoken==0.5.2  # Optionnel, nécessite Rust compiler

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
